feat : move Macros from GL3 to Core namespace
