+++ Sources/GL3/DebugUtils.cpp
@@ -1,5 +1,5 @@
 #include <GL3/DebugUtils.hpp>
-#include <GL3/Macros.hpp>
+#include <Core/Macros.hpp>
 #include <iostream>
 #include <cstdio>
